pyproject.toml
[tool.poetry]
name = "playlist-app"
version = "0.1.0"
description = ""
authors = ["Sarah <vanwars@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
requests = "^2.31.0"
spotipy = "^2.23.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
